adding K8s scripts to ensure elasticsearch has properly started
On a rolling restart strategy in kubernetes, it doesn't ensure
elasticsearch has proprely finish to boot.

For example if there is a bad value in the configuration file, it will
start booting, kubernetes doesn't wait until the port is open, then
crash and finally you've got a full failed cluster because Kubernetes
will continue to rollout without checking.

Using a postStart hook ensure the port is open until it restarts the
next instance. The script could also be added and mounted in a configmap
but I think it's preferable to add it in the container with all other
potential useful scripts for kubernetes.