run.sh: retry downloading plugins upon failure
In certain conditions, the network may not entirely
ready when the container starts. This prevents the plugins
from being downloaded, which may not be desirable as
liveness probes won't pick up the fact that ES does not run
all the features that were expected to be present (e.g.
prometheus metrics exporting).

Instead, this commit proposes retries upon installation
failures. In Kubernetes, the pod will eventually get killed
and restarted by the liveness probs after the initialDelaySeconds
delay has expired.